requirements.txt
aiohttp==3.8.5
deepmultilingualpunctuation==1.0.1
Flask==2.3.2
germansentiment==1.1.0
Logbook==1.6.0
matplotlib==3.7.2
moviepy==1.0.3
nltk==3.8.1
python-dotenv==1.0.0
Requests==2.31.0
simple_youtube_api==0.2.8
tabulate==0.9.0
tqdm==4.65.0
tweepy==4.14.0
vosk==0.3.45
